Remove old pylons support for CKAN < 2.9
