--- ckanext/validation/plugin/pylons_plugin.py
@@ -1,18 +1,0 @@
-from ckan import plugins as p
-
-
-class ValidationMixin(p.SingletonPlugin):
-    p.implements(p.IRoutes, inherit=True)
-
-    # IRoutes
-
-    def before_map(self, map_):
-
-        controller = u'ckanext.validation.controller:ValidationController'
-
-        map_.connect(
-            u'validation_read',
-            u'/dataset/{id}/resource/{resource_id}/validation',
-            controller=controller, action=u'validation')
-
-        return map_
